backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
google-genai
PyMuPDF
python-docx
python-jose[cryptography]
python-multipart
pydantic
pydantic-settings
slowapi
pinecone
